refactor(scheduler): try move from interpreter to schedulers
