+++ src/query/service/src/schedulers/fragments/fragmenter.rs
@@ -23,8 +23,8 @@ use crate::api::DataExchange;
 use crate::api::MergeExchange;
 use crate::api::ShuffleDataExchange;
 use crate::clusters::ClusterHelper;
-use crate::interpreters::fragments::plan_fragment::FragmentType;
-use crate::interpreters::fragments::PlanFragment;
+use crate::schedulers::fragments::plan_fragment::FragmentType;
+use crate::schedulers::PlanFragment;
 use crate::sessions::QueryContext;
 use crate::sql::executor::Exchange;
 use crate::sql::executor::ExchangeSink;
